Split workflow CLI target helpers
