+++ gestalt/cli/src/commands/mod.rs
@@ -8,4 +8,5 @@ pub mod invoke;
 mod plugin_errors;
 pub mod plugins;
 pub mod tokens;
+mod workflow_target;
 pub mod workflows;
