+++ gestalt/cli/src/commands/workflows.rs
@@ -9,13 +9,15 @@ use crate::cli::{
 use crate::output::{self, Format};
 use crate::params::{self, ParamEntry};
 
+use super::workflow_target::{
+    PluginTargetUpdate, build_plugin_target, literal_input, merge_plugin_target_flags,
+    target_has_plugin, target_operation, target_plugin,
+};
+
 const EVENTS_PATH: &str = "/api/v1/workflow/events";
 const SCHEDULES_PATH: &str = "/api/v1/workflow/schedules";
 const TRIGGERS_PATH: &str = "/api/v1/workflow/event-triggers";
 const RUNS_PATH: &str = "/api/v1/workflow/runs";
-
-type JsonMap = Map<String, Value>;
-type SinglePluginStep<'a> = (&'a JsonMap, &'a JsonMap);
 
 pub fn list(client: &ApiClient, plugin: Option<&str>, format: Format) -> Result<()> {
     let resp = client
@@ -36,7 +38,7 @@ pub fn get(client: &ApiClient, id: &str, format: Format) -> Result<()> {
 
 pub fn create(client: &ApiClient, args: &WorkflowScheduleCreateArgs, format: Format) -> Result<()> {
     let input = build_optional_map(&args.params, args.input_file.as_deref())?;
-    let input_value = input.as_ref().map(workflow_literal_input);
+    let input_value = input.as_ref().map(literal_input);
     let body = build_upsert_body(
         &args.cron,
         args.timezone.as_deref(),
@@ -378,7 +380,7 @@ fn build_upsert_body(
     build_schedule_upsert_body(
         cron,
         timezone,
-        build_target_object(plugin, operation, connection, instance, input),
+        build_plugin_target(plugin, operation, connection, instance, input),
         paused,
     )
 }
@@ -453,8 +455,8 @@ fn build_trigger_target(args: &WorkflowTriggerCreateArgs) -> Result<Value> {
         .as_deref()
         .ok_or_else(|| anyhow!("workflow trigger plugin target requires --operation"))?;
     let input = build_optional_map(&args.params, args.input_file.as_deref())?;
-    let input_value = input.as_ref().map(workflow_literal_input);
-    Ok(build_target_object(
+    let input_value = input.as_ref().map(literal_input);
+    Ok(build_plugin_target(
         plugin,
         operation,
         args.connection.as_deref(),
@@ -474,64 +476,7 @@ fn merge_update(args: &WorkflowScheduleUpdateArgs, existing: &Value) -> Result<V
 
     let timezone = resolve_optional_string(args.timezone.as_deref(), existing["timezone"].as_str());
 
-    let target = if schedule_target_overrides(args) {
-        let (step, plugin_step) = target_single_plugin_step(existing).ok_or_else(|| {
-            anyhow!(
-                "cannot apply plugin target flags to an existing non-plugin or multi-step schedule; recreate it with a full target definition"
-            )
-        })?;
-        let plugin = match args.plugin.as_deref() {
-            Some(value) => value.to_string(),
-            None => plugin_step
-                .get("name")
-                .and_then(Value::as_str)
-                .ok_or_else(|| {
-                    anyhow!("existing schedule is missing target.steps plugin name; pass --plugin")
-                })?
-                .to_string(),
-        };
-        let operation = match args.operation.as_deref() {
-            Some(value) => value.to_string(),
-            None => plugin_step
-                .get("operation")
-                .and_then(Value::as_str)
-                .ok_or_else(|| {
-                    anyhow!(
-                        "existing schedule is missing target.steps plugin operation; pass --operation"
-                    )
-                })?
-                .to_string(),
-        };
-        let connection = resolve_optional_string(
-            args.connection.as_deref(),
-            plugin_step.get("connection").and_then(Value::as_str),
-        );
-        let instance = resolve_optional_string(
-            args.instance.as_deref(),
-            plugin_step.get("instance").and_then(Value::as_str),
-        );
-        let input = if args.clear_input {
-            None
-        } else if !args.params.is_empty() || args.input_file.is_some() {
-            build_optional_map(&args.params, args.input_file.as_deref())?
-                .map(|input| workflow_literal_input(&input))
-        } else {
-            plugin_step.get("input").cloned()
-        };
-        build_target_object_from_step(
-            Some(step),
-            &plugin,
-            &operation,
-            connection.as_deref(),
-            instance.as_deref(),
-            input.as_ref(),
-        )
-    } else {
-        existing
-            .get("target")
-            .cloned()
-            .ok_or_else(|| anyhow!("existing schedule is missing target"))?
-    };
+    let target = merge_schedule_target_update(args, existing)?;
 
     let paused = if args.paused {
         true
@@ -563,64 +508,7 @@ fn merge_trigger_update(args: &WorkflowTriggerUpdateArgs, existing: &Value) -> R
         args.subject.as_deref(),
         existing["match"]["subject"].as_str(),
     );
-    let target = if trigger_target_overrides(args) {
-        let (step, plugin_step) = target_single_plugin_step(existing).ok_or_else(|| {
-            anyhow!(
-                "cannot apply plugin target flags to an existing non-plugin or multi-step trigger; recreate it with a full target definition"
-            )
-        })?;
-        let plugin = match args.plugin.as_deref() {
-            Some(value) => value.to_string(),
-            None => plugin_step
-                .get("name")
-                .and_then(Value::as_str)
-                .ok_or_else(|| {
-                    anyhow!("existing trigger is missing target.steps plugin name; pass --plugin")
-                })?
-                .to_string(),
-        };
-        let operation = match args.operation.as_deref() {
-            Some(value) => value.to_string(),
-            None => plugin_step
-                .get("operation")
-                .and_then(Value::as_str)
-                .ok_or_else(|| {
-                    anyhow!(
-                        "existing trigger is missing target.steps plugin operation; pass --operation"
-                    )
-                })?
-                .to_string(),
-        };
-        let connection = resolve_optional_string(
-            args.connection.as_deref(),
-            plugin_step.get("connection").and_then(Value::as_str),
-        );
-        let instance = resolve_optional_string(
-            args.instance.as_deref(),
-            plugin_step.get("instance").and_then(Value::as_str),
-        );
-        let input = if args.clear_input {
-            None
-        } else if !args.params.is_empty() || args.input_file.is_some() {
-            build_optional_map(&args.params, args.input_file.as_deref())?
-                .map(|input| workflow_literal_input(&input))
-        } else {
-            plugin_step.get("input").cloned()
-        };
-        build_target_object_from_step(
-            Some(step),
-            &plugin,
-            &operation,
-            connection.as_deref(),
-            instance.as_deref(),
-            input.as_ref(),
-        )
-    } else {
-        existing
-            .get("target")
-            .cloned()
-            .ok_or_else(|| anyhow!("existing trigger is missing target"))?
-    };
+    let target = merge_trigger_target_update(args, existing)?;
 
     let paused = if args.paused {
         true
@@ -640,24 +528,57 @@ fn merge_trigger_update(args: &WorkflowTriggerUpdateArgs, existing: &Value) -> R
     ))
 }
 
-fn schedule_target_overrides(args: &WorkflowScheduleUpdateArgs) -> bool {
-    args.plugin.is_some()
-        || args.operation.is_some()
-        || args.connection.is_some()
-        || args.instance.is_some()
-        || args.clear_input
-        || !args.params.is_empty()
-        || args.input_file.is_some()
+fn merge_schedule_target_update(
+    args: &WorkflowScheduleUpdateArgs,
+    existing: &Value,
+) -> Result<Value> {
+    let (input, replace_input) = target_input_update(&args.params, args.input_file.as_deref())?;
+    merge_plugin_target_flags(
+        existing,
+        PluginTargetUpdate {
+            resource: "schedule",
+            plugin: args.plugin.as_deref(),
+            operation: args.operation.as_deref(),
+            connection: args.connection.as_deref(),
+            instance: args.instance.as_deref(),
+            clear_input: args.clear_input,
+            replace_input,
+            input: input.as_ref(),
+        },
+    )
 }
 
-fn trigger_target_overrides(args: &WorkflowTriggerUpdateArgs) -> bool {
-    args.plugin.is_some()
-        || args.operation.is_some()
-        || args.connection.is_some()
-        || args.instance.is_some()
-        || args.clear_input
-        || !args.params.is_empty()
-        || args.input_file.is_some()
+fn merge_trigger_target_update(
+    args: &WorkflowTriggerUpdateArgs,
+    existing: &Value,
+) -> Result<Value> {
+    let (input, replace_input) = target_input_update(&args.params, args.input_file.as_deref())?;
+    merge_plugin_target_flags(
+        existing,
+        PluginTargetUpdate {
+            resource: "trigger",
+            plugin: args.plugin.as_deref(),
+            operation: args.operation.as_deref(),
+            connection: args.connection.as_deref(),
+            instance: args.instance.as_deref(),
+            clear_input: args.clear_input,
+            replace_input,
+            input: input.as_ref(),
+        },
+    )
+}
+
+fn target_input_update(
+    params: &[ParamEntry],
+    input_file: Option<&str>,
+) -> Result<(Option<Value>, bool)> {
+    let replace_input = !params.is_empty() || input_file.is_some();
+    let input = if replace_input {
+        build_optional_map(params, input_file)?.map(|input| literal_input(&input))
+    } else {
+        None
+    };
+    Ok((input, replace_input))
 }
 
 fn resolve_optional_string(arg: Option<&str>, existing: Option<&str>) -> Option<String> {
@@ -666,141 +587,6 @@ fn resolve_optional_string(arg: Option<&str>, existing: Option<&str>) -> Option<
         Some(value) => Some(value.to_string()),
         None => existing.map(str::to_string),
     }
-}
-
-fn build_target_object(
-    plugin: &str,
-    operation: &str,
-    connection: Option<&str>,
-    instance: Option<&str>,
-    input: Option<&Value>,
-) -> Value {
-    build_target_object_from_step(None, plugin, operation, connection, instance, input)
-}
-
-fn build_target_object_from_step(
-    existing_step: Option<&Map<String, Value>>,
-    plugin: &str,
-    operation: &str,
-    connection: Option<&str>,
-    instance: Option<&str>,
-    input: Option<&Value>,
-) -> Value {
-    let mut step = existing_step.cloned().unwrap_or_default();
-    let mut plugin_target = existing_step
-        .and_then(|step| step.get("plugin"))
-        .and_then(Value::as_object)
-        .cloned()
-        .unwrap_or_default();
-    plugin_target.insert("name".to_string(), Value::String(plugin.to_string()));
-    plugin_target.insert(
-        "operation".to_string(),
-        Value::String(operation.to_string()),
-    );
-    if let Some(connection) = connection {
-        plugin_target.insert(
-            "connection".to_string(),
-            Value::String(connection.to_string()),
-        );
-    } else {
-        plugin_target.remove("connection");
-    }
-    if let Some(instance) = instance {
-        plugin_target.insert("instance".to_string(), Value::String(instance.to_string()));
-    } else {
-        plugin_target.remove("instance");
-    }
-    if let Some(input) = input {
-        plugin_target.insert("input".to_string(), input.clone());
-    } else {
-        plugin_target.remove("input");
-    }
-    if !matches!(step.get("id"), Some(Value::String(value)) if !value.is_empty()) {
-        step.insert("id".to_string(), Value::String(operation.to_string()));
-    }
-    step.remove("agent");
-    step.insert("plugin".to_string(), Value::Object(plugin_target));
-    let mut target = Map::new();
-    target.insert("steps".to_string(), Value::Array(vec![Value::Object(step)]));
-    Value::Object(target)
-}
-
-fn workflow_literal_input(input: &Map<String, Value>) -> Value {
-    let mut value = Map::new();
-    value.insert("literal".to_string(), Value::Object(input.clone()));
-    Value::Object(value)
-}
-
-fn target_plugin<'a>(value: &'a Value, preferred_plugin: Option<&str>) -> Option<&'a str> {
-    target_plugin_step(value, preferred_plugin)?
-        .get("name")
-        .and_then(Value::as_str)
-}
-
-fn target_operation<'a>(value: &'a Value, preferred_plugin: Option<&str>) -> Option<&'a str> {
-    target_plugin_field(value, preferred_plugin, "operation")
-}
-
-fn target_single_plugin_step(value: &Value) -> Option<SinglePluginStep<'_>> {
-    let steps = value.get("target")?.get("steps")?.as_array()?;
-    if steps.len() != 1 {
-        return None;
-    }
-    let step = steps.first()?.as_object()?;
-    if step.get("agent").is_some() {
-        return None;
-    }
-    let plugin = step.get("plugin").and_then(Value::as_object)?;
-    Some((step, plugin))
-}
-
-fn target_has_plugin(value: &Value, plugin: &str) -> bool {
-    value
-        .get("target")
-        .and_then(|target| target.get("steps"))
-        .and_then(Value::as_array)
-        .map(|steps| {
-            steps.iter().any(|step| {
-                step.get("plugin")
-                    .and_then(Value::as_object)
-                    .and_then(|plugin_step| plugin_step.get("name"))
-                    .and_then(Value::as_str)
-                    == Some(plugin)
-            })
-        })
-        .unwrap_or(false)
-}
-
-fn target_plugin_step<'a>(
-    value: &'a Value,
-    preferred_plugin: Option<&str>,
-) -> Option<&'a Map<String, Value>> {
-    let steps = value.get("target")?.get("steps")?.as_array()?;
-    let mut first_plugin_step = None;
-    for step in steps {
-        let Some(plugin_step) = step.get("plugin").and_then(Value::as_object) else {
-            continue;
-        };
-        if first_plugin_step.is_none() {
-            first_plugin_step = Some(plugin_step);
-        }
-        if preferred_plugin.is_some_and(|preferred_plugin| {
-            plugin_step.get("name").and_then(Value::as_str) == Some(preferred_plugin)
-        }) {
-            return Some(plugin_step);
-        }
-    }
-    first_plugin_step
-}
-
-fn target_plugin_field<'a>(
-    value: &'a Value,
-    preferred_plugin: Option<&str>,
-    field: &str,
-) -> Option<&'a str> {
-    target_plugin_step(value, preferred_plugin)?
-        .get(field)
-        .and_then(Value::as_str)
 }
 
 fn build_event_match(event_type: &str, source: Option<&str>, subject: Option<&str>) -> Value {
@@ -1049,41 +835,5 @@ fn format_bool(value: Option<bool>) -> String {
         Some(true) => "yes".to_string(),
         Some(false) => "no".to_string(),
         None => "-".to_string(),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn target_plugin_fields_prefer_matching_plugin_step() {
-        let value = json!({
-            "target": {
-                "steps": [
-                    {
-                        "id": "first",
-                        "plugin": {
-                            "name": "github",
-                            "operation": "createIssue"
-                        }
-                    },
-                    {
-                        "id": "second",
-                        "plugin": {
-                            "name": "slack",
-                            "operation": "reply"
-                        }
-                    }
-                ]
-            }
-        });
-
-        assert_eq!(target_plugin(&value, None), Some("github"));
-        assert_eq!(target_operation(&value, None), Some("createIssue"));
-        assert_eq!(target_plugin(&value, Some("slack")), Some("slack"));
-        assert_eq!(target_operation(&value, Some("slack")), Some("reply"));
-        assert_eq!(target_plugin(&value, Some("jira")), Some("github"));
-        assert_eq!(target_operation(&value, Some("jira")), Some("createIssue"));
     }
 }
